+++ src/webview/types.ts
@@ -1,0 +1,13 @@
+export interface ProcessInfo {
+  name: string;
+  type: "process" | "subworkflow";
+  hasTest: boolean;
+  testFilePath?: string;
+}
+
+export interface PipelineNode {
+  fileName: string;
+  filePath: string;
+  processes: ProcessInfo[];
+  children: PipelineNode[];
+}
